Add "Read More" functionality to the about page.
This commit adds the "Read More" functionality to the about page. It includes the addition of a new partial file, "read_more.html", which contains the additional content that is revealed when the "Read More" button is clicked. This functionality allows for more detailed information to be displayed, enhancing the user experience.